Streamlit_App/requirements.txt
newspaper3k
spacy
lxml_html_clean